+++ checkpoint/orbax/checkpoint/experimental/v1/_src/loading/layout_loading_test.py
@@ -75,9 +75,9 @@ class LayoutLoadingTest(parameterized.TestCase):
     )
 
   def test_load_safetensors_checkpoint(self):
-    with context_lib.Context(
-        checkpoint_layout=options_lib.CheckpointLayout.SAFETENSORS
-    ):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = options_lib.CheckpointLayout.SAFETENSORS
+    with ctx:
       pytree = loading.load_pytree(self.safetensors_path)
     self.assertIsInstance(pytree, dict)
     np.testing.assert_array_equal(pytree['a'], self.object_to_save['a'])
@@ -97,7 +97,9 @@ class LayoutLoadingTest(parameterized.TestCase):
   )
   def test_load_bad_path_orbax_ckpt(self, layout_enum):
     # User provides a directory of Orbax checkpoints, not specific one.
-    with context_lib.Context(checkpoint_layout=layout_enum):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = layout_enum
+    with ctx:
       with self.assertRaises(InvalidLayoutError):
         loading.load_pytree(
             epath.Path(self.test_dir.full_path),
@@ -108,7 +110,9 @@ class LayoutLoadingTest(parameterized.TestCase):
   )
   def test_load_bad_path_safetensors_ckpt(self, layout_enum):
     # User provides a empty directory of SafeTensors checkpoints, not a file.
-    with context_lib.Context(checkpoint_layout=layout_enum):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = layout_enum
+    with ctx:
       with self.assertRaises(InvalidLayoutError):
         loading.load_pytree(
             epath.Path(self.test_dir_safetensors.full_path),
@@ -118,9 +122,9 @@ class LayoutLoadingTest(parameterized.TestCase):
     safetensors_dir = epath.Path(self.test_dir_safetensors.full_path)
     safetensors_path = safetensors_dir / 'model.safetensors'
     np_save_file(self.object_to_save, safetensors_path)
-    with context_lib.Context(
-        checkpoint_layout=options_lib.CheckpointLayout.SAFETENSORS
-    ):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = options_lib.CheckpointLayout.SAFETENSORS
+    with ctx:
       pytree = loading.load_pytree(safetensors_dir)
     self.assertIsInstance(pytree, dict)
     np.testing.assert_array_equal(pytree['a'], self.object_to_save['a'])
@@ -180,7 +184,9 @@ class LayoutLoadingTest(parameterized.TestCase):
     else:
       directory = self.orbax_pytree_path
 
-    with context_lib.Context(checkpoint_layout=layout):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = layout
+    with ctx:
       if layout != options_lib.CheckpointLayout.SAFETENSORS:
         with self.assertRaises(NotImplementedError):
           loading.load_pytree_async(directory)
@@ -197,9 +203,9 @@ class LayoutLoadingTest(parameterized.TestCase):
   # TODO(b/431045454): Add tests for abstract_checkpointables.
 
   def test_load_auto_resolution_mode_orbax(self):
-    with context_lib.Context(
-        checkpoint_layout=options_lib.CheckpointLayout.ORBAX
-    ):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = options_lib.CheckpointLayout.ORBAX
+    with ctx:
       loaded_orbax = loading.load_pytree(
           self.orbax_pytree_path,
           checkpointable_name=checkpoint_layout.AUTO_CHECKPOINTABLE_KEY,
@@ -207,9 +213,9 @@ class LayoutLoadingTest(parameterized.TestCase):
     test_utils.assert_tree_equal(self, self.object_to_save, loaded_orbax)
 
   def test_load_auto_resolution_mode_safetensors(self):
-    with context_lib.Context(
-        checkpoint_layout=options_lib.CheckpointLayout.SAFETENSORS
-    ):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = options_lib.CheckpointLayout.SAFETENSORS
+    with ctx:
       loaded_safe = loading.load_pytree(
           self.safetensors_path,
           checkpointable_name=checkpoint_layout.AUTO_CHECKPOINTABLE_KEY,
@@ -227,9 +233,9 @@ class LayoutLoadingTest(parameterized.TestCase):
     saving.save_checkpointables(multiple_path, checkpointables)
 
     # Triggering AUTO loading mode should prioritize resolving 'pytree'.
-    with context_lib.Context(
-        checkpoint_layout=options_lib.CheckpointLayout.ORBAX
-    ):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = options_lib.CheckpointLayout.ORBAX
+    with ctx:
       loaded = loading.load_pytree(multiple_path)
 
     test_utils.assert_tree_equal(self, checkpointables['pytree'], loaded)
@@ -242,9 +248,9 @@ class LayoutLoadingTest(parameterized.TestCase):
     fallback_path = epath.Path(self.test_dir.full_path) / 'fallback_checkpoint'
     saving.save_checkpointables(fallback_path, custom_checkpointables)
 
-    with context_lib.Context(
-        checkpoint_layout=options_lib.CheckpointLayout.ORBAX
-    ):
+    ctx = context_lib.Context()
+    ctx.checkpoint_layout = options_lib.CheckpointLayout.ORBAX
+    with ctx:
       loaded = loading.load_pytree(
           fallback_path,
           checkpointable_name=checkpoint_layout.AUTO_CHECKPOINTABLE_KEY,