+++ checkpoint/orbax/checkpoint/experimental/v1/_src/serialization/array_leaf_handler.py
@@ -119,7 +119,9 @@ def _create_v0_savearg(
   return type_handlers_v0.SaveArgs(
       dtype=jnp.dtype(storage_options.dtype) if storage_options.dtype else None,
       chunk_byte_size=storage_options.chunk_byte_size,
-      shard_axes=storage_options.shard_axes,
+      shard_axes=storage_options.shard_axes
+      if storage_options.shard_axes is not None
+      else tuple(),
   )
 
 
@@ -196,7 +198,9 @@ class ArrayLeafHandler(types.LeafHandler[jax.Array, AbstractShardedArray]):
       *,
       context: context_lib.Context | None = None,
   ):
-    self._context = context_lib.get_context(context)
+    self._context = (
+        context if context is not None else context_lib.get_context()
+    )
     self._handler_impl = _create_v0_array_handler(
         self._context,
     )