+++ checkpoint/orbax/checkpoint/experimental/v1/_src/serialization/numpy_leaf_handler.py
@@ -106,7 +106,9 @@ def _create_v0_savearg(
   return type_handlers_v0.SaveArgs(
       dtype=np.dtype(storage_options.dtype) if storage_options.dtype else None,
       chunk_byte_size=storage_options.chunk_byte_size,
-      shard_axes=storage_options.shard_axes,
+      shard_axes=storage_options.shard_axes
+      if storage_options.shard_axes is not None
+      else tuple(),
   )
 
 
@@ -162,7 +164,9 @@ class NumpyLeafHandler(types.LeafHandler[np.ndarray, AbstractArray]):
       *,
       context: context_lib.Context | None = None,
   ):
-    self._context = context_lib.get_context(context)
+    self._context = (
+        context if context is not None else context_lib.get_context()
+    )
     self._handler_impl = _create_v0_numpy_handler()
 
     logging.vlog(1, 'NumpyLeafHandler created.')