+++ checkpoint/orbax/checkpoint/experimental/v1/_src/testing/compatibility/pytree_metadata_compatibility_test_base.py
@@ -186,11 +186,9 @@ class PytreeMetadataCompatibilityTestBase(parameterized.TestCase):
         is_pytree,
     )
 
-    with ocp.Context(
-        checkpointables_options=ocp.options.CheckpointablesOptions(
-            registry=registry
-        )
-    ):
+    ctx = ocp.Context()
+    ctx.checkpointables.registry = registry
+    with ctx:
       if error_type is None:
         loaded = ocp.pytree_metadata(
             path,